Antispam semi-implementado. Borda nas letras. Não da mais para casar com sí mesmo.

+++ src/main/java/com/kuuhaku/command/commands/misc/SayCommand.java
@@ -3,6 +3,7 @@ package com.kuuhaku.command.commands.misc;
 import com.kuuhaku.command.Category;
 import com.kuuhaku.command.Command;
 import com.kuuhaku.utils.Helper;
+import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.*;
 import net.dv8tion.jda.core.events.Event;
 
@@ -17,7 +18,7 @@ public class SayCommand extends Command {
 		
 		if(args.length == 0) { channel.sendMessage(":x: | Você precisa definir uma mensagem.").queue(); return; }
 
-		message.delete().queue();
+		if (guild.getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)) message.delete().queue();
 		channel.sendMessage(Helper.makeEmoteFromMention(args)).queue();
 	}
